+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Optional Environment variable AMQP_PREFETCH_COUNT
 
+### Changed
+- Renamed env var REDIS_URI to REDIS_CONFIGURATION_STRING (same functionality)
+
 
 ## [0.4.0] - 2020-03-04
 ### Added